+++ src/main/java/com/springboot/app/EjercicioSpring/repository/EmployeesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.springboot.app.EjercicioSpring.model.Employees;
 
 @Repository
-public interface EmployeesRepository extends JpaRepository<Employees, Long> {
+public interface EmployeesRepository extends JpaRepository<Employees, Integer> {
 
 }
