+++ src/main/java/com/springboot/app/EjercicioSpring/service/AirportsService.java
@@ -8,13 +8,13 @@ import com.springboot.app.EjercicioSpring.model.Airports;
 public interface AirportsService {
 	
 	//Obtener por id => FindById
-	Airports getAirportsById(long airportId);
+	Airports getAirportsById(int airportId);
 	
 	//Obtener todo => FindAl
 		List<Airports> getAllAirports();
 	
 	//Eliminar => Delete
-	void deleteAirports(long id);
+	void deleteAirports(int id);
 
 	//Actualizar => Update
 	Airports updateAirports(Airports airport);
