+++ src/main/java/com/springboot/app/EjercicioSpring/service/EmployeesService.java
@@ -8,13 +8,13 @@ import com.springboot.app.EjercicioSpring.model.Employees;
 public interface EmployeesService {
 	
 	//Obtener por id => FindById
-	Employees getEmployeesById(long employeeId);
+	Employees getEmployeesById(int employeeId);
 			
 	//Obtener todo => FindAl
 	List<Employees> getAllEmployees();
 			
 	//Eliminar => Delete
-	void deleteEmployees(long id);
+	void deleteEmployees(int id);
 
 	//Actualizar => Update
 	Employees updateEmployees(Employees employee);
